% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{build_graphql_query}
\alias{build_graphql_query}
\title{Builds a query for use in \code{get_linear_issues}}
\usage{
build_graphql_query(team_id, cursor = NULL, year = NULL)
}
\arguments{
\item{team_id}{The id for the team. Use get_linear_teams() to find the id of the team you want.}

\item{cursor}{The id of the \code{after} cursor, obtained with \code{page_info$endCursor}.}

\item{year}{If specified, all the issues (except the archived issues) since that year are returned. Default is \code{NULL} that returns all the issues to date, except the archived issues.}
}
\value{
character
}
\description{
Builds a query for use in \code{get_linear_issues}
}
