% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_linear_issues}
\alias{get_linear_issues}
\title{Returns a tibble with the issues for a given team_id}
\usage{
get_linear_issues(team_id, api_url = "https://api.linear.app/graphql")
}
\arguments{
\item{team_id}{The id for the team. Use get_linear_teams() to find the id of the team you want.}

\item{api_url}{Defaults to the current linear api url, but can be changed}
}
\value{
tibble
}
\description{
Returns a tibble with the issues for a given team_id
}
