% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_linear_api_request}
\alias{make_linear_api_request}
\title{Make an API request}
\usage{
make_linear_api_request(
  request_string,
  api_url = "https://api.linear.app/graphql"
)
}
\arguments{
\item{request_string}{GraphQL API query.}

\item{api_url}{Defaults to the current linear api url, but can be changed.}
}
\value{
httr2 response
}
\description{
Make an API request
}
