% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{me}
\alias{me}
\title{Returns a tibble with your id, name, and email. Useful for checking if you have successfully authenticated with the linear API.}
\usage{
me()
}
\value{
tibble
}
\description{
Returns a tibble with your id, name, and email. Useful for checking if you have successfully authenticated with the linear API.
}
