% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{store_linear_api_key}
\alias{store_linear_api_key}
\title{Stores your API key in your keyring.}
\usage{
store_linear_api_key(api_key)
}
\arguments{
\item{api_key}{Your personal linear API key. See README.md for information.}
}
\value{
nothing
}
\description{
Stores your API key in your keyring.
}
