% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linearity.R
\name{check_linearity}
\alias{check_linearity}
\title{Check Linearity Assumption}
\usage{
check_linearity(object, use_test = TRUE)
}
\arguments{
\item{object}{A \code{clean_lm} object returned by \code{clean_lm()}.}

\item{use_test}{Logical; if TRUE, performs Ramsey RESET test using \code{lmtest::resettest()}.}
}
\value{
A list containing:
\itemize{
 \item \code{plot}: ggplot object showing residuals vs. fitted values
 \item \code{test_result}: result of RESET test (if \code{use_test = TRUE})
 \item \code{linearity_ok}: logical flag indicating whether linearity assumption holds
}
}
\description{
Assesses linearity assumption using a residuals vs. fitted
values plot and Ramsey RESET test.
}
\examples{
data <- data.frame(x = 1:100, y = 3 * (1:100) + rnorm(100))
fit <- clean_lm(y ~ x, data)
check_linearity(fit)


}
