% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lm.R
\name{clean_lm}
\alias{clean_lm}
\title{Clean Linear Model
Fits a linear model and stores data for diagnostics.}
\usage{
clean_lm(formula, data)
}
\arguments{
\item{formula}{A formula to fit the data, like 'y~x'}

\item{data}{A data frame containing the variables fitted in the formula}
}
\value{
A list containing the fitted model and residuals
}
\description{
Clean Linear Model
Fits a linear model and stores data for diagnostics.
}
\examples{
data<- data.frame(x= 1:10, y = 2*(1:10) + rnorm(10))
result <- clean_lm(y~x, data)
summary(result$model)
}
