% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{create_linear_issue}
\alias{create_linear_issue}
\title{Creates a new issue within a team specified using the team's id.}
\usage{
create_linear_issue(title, description, team_id)
}
\arguments{
\item{title}{The issue title}

\item{description}{The issue description}

\item{team_id}{The id for the team. Use get_linear_teams() to find the id of the team you want.}
}
\value{
tibble
}
\description{
Creates a new issue within a team specified using the team's id.
}
