% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_linear_current_user}
\alias{get_linear_current_user}
\title{Returns a tibble with your id, name, and email. Useful for checking if you have successfully authenticated with the linear API.}
\usage{
get_linear_current_user()
}
\value{
tibble
}
\description{
Returns a tibble with your id, name, and email. Useful for checking if you have successfully authenticated with the linear API.
}
